lanchonete/project/backend/pyproject.toml
[tool.poetry]
name = "lanchonete-api"
version = "0.1.0"
description = "Sistema de pedidos para lanchonete"
authors = ["Dev"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.3.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.4"
httpx = "^0.27.0"
openai = "^1.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
